refactor(domain, dto): 멤버 도메인 및 패키지 구조 변경
도메인 모델과 패키지 구조를 변경하여 프로젝트의 모듈화를 개선하고, 코드의 가독성을 향상시켰습니다. 주요 변경 사항은 다음과 같습니다:

1. 멤버 관련 도메인 클래스들의 패키지 경로를 `sidepair.member.domain`에서 `sidepair.domain.member`로 이동시켰습니다. 이는 도메인 중심의 설계를 강화하고, 관련 도메인 모델들을 더 직관적으로 관리하기 위함입니다.

2. 멤버 관련 DTO들의 패키지 경로를 `sidepair.member.configuration.dto`에서 `sidepair.service.dto.mamber`로 변경하였습니다. 오타가 있던 `mamber`를 `member`로 수정하여 일관성을 유지하고, 서비스 레이어와의 연관성을 명확히 했습니다.

3. 멤버 도메인과 DTO에서 사용되는 클래스 및 인터페이스의 네이밍을 보다 명확하게 변경하여, 역할과 책임을 쉽게 파악할 수 있도록 하였습니다.

4. `Member` 엔티티 내 `EncryptedPassword`, `MemberSkills`, `MemberProfile` 등의 컴포넌트들을 재정의하여, 멤버 도메인의 캡슐화와 정보 은닉을 강화하였습니다.

이러한 구조적 변경은 팀원들이 프로젝트의 아키텍처를 더 쉽게 이해하고, 향후 유지보수 및 기능 확장을 용이하게 할 것으로 기대됩니다.
