+++ src/main/java/sidepair/domain/member/vo/Email.java
@@ -1,11 +1,11 @@
-package sidepair.member.domain.vo;
+package sidepair.domain.member.vo;
 
 import jakarta.persistence.Column;
 import java.util.Objects;
 import java.util.regex.Pattern;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import sidepair.member.exception.MemberException;
+import sidepair.domain.member.exception.MemberException;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Email {
