+++ src/main/java/sidepair/service/dto/mamber/response/MemberInformationForPublicResponse.java
@@ -1,4 +1,4 @@
-package sidepair.member.configuration.response;
+package sidepair.service.dto.mamber.response;
 
 import java.util.List;
